Backend/python/requirements.txt
pytesseract
Pillow
groq
easyocr
python-dotenv
torch
torchvision
torchaudio
python-multipart

